requirements.txt
coverage
twine
wheel
pytest
pytest-cov
pytest-djangoapp